requirements.txt
amqp==5.3.1
billiard==4.2.1
blinker==1.9.0
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.7.0
Flask==3.1.0
flask-cors==5.0.1
Flask-PyMongo==3.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.0
MarkupSafe==3.0.2
numpy==2.2.3
pandas==2.2.3
prompt_toolkit==3.0.50
py4j==0.10.9.5
pymongo==4.11.2
pyspark==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
redis==5.2.1
setuptools==76.0.0
six==1.17.0
tzdata==2025.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
